adopt PR template for DCO/CLA-like agreement
This is after publishing an article with my own viewpoints:

https://devlog.hexops.com/2021/cla-not-for-open-source-use-a-developer-certificate-of-origin

Having it called out by an actual professional lawyer I look up to, Kyle E. Mitchell, as spreading misinformation:

https://lobste.rs/s/lvmb5i/clas_are_not_for_open_source_use_developer#c_n2k3zd

Reading Ben Balter's article and viewpoints (GitHub's Chief of Staff for Security):

https://ben.balter.com/2018/01/02/why-you-probably-shouldnt-add-a-cla-to-your-open-source-project/

Reading and doing my best to understand Kyle's viewpoints in:

https://writing.kemitchell.com/2018/01/06/CLAs-Are-Not-a-Sham.html

And coming to a sort of middle-ground appreciation for both viewpoints,
with one hand being leaning towards Ben's viewpoints thinking 'maybe GitHub's ToS inbound=outbound
viewpoint is semi-ok' and the other hand being towards Kyle thinking 'that's very out-of-sight-out-of-mind,
non-explicit and maybe CLAs aren't so bad after all, but the current format of them is'.

This is a "it's definitely better than nothing", "hopefully better than a DCO in terms of legal guards due
to its explicitness", and "definitely a better developer experience than a DCO/CLA in either case" approach.

Again, I am not a lawyer. This isn't legal advice, it's just my best interpretation of what the
best thing to do is as I continue to learn about software law.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>